requirements.txt
loguru
mysql-replication
pymysql